Modified MainSignInFragmentEnterpriseFlowTest. Removed CreatePrivateKeyFlowEnterpriseFlowTest as deprecated.| #1947
